+++ api/global/internal/meter_test.go
@@ -22,8 +22,6 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"go.opentelemetry.io/otel/api/kv/value"
-
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel/api/global"
@@ -39,7 +37,7 @@ type measured struct {
 	Name                   string
 	InstrumentationName    string
 	InstrumentationVersion string
-	Labels                 map[kv.Key]value.Value
+	Labels                 map[kv.Key]kv.Value
 	Number                 metric.Number
 }
 
@@ -59,8 +57,8 @@ func asStructs(batches []metrictest.Batch) []measured {
 	return r
 }
 
-func asMap(kvs ...kv.KeyValue) map[kv.Key]value.Value {
-	m := map[kv.Key]value.Value{}
+func asMap(kvs ...kv.KeyValue) map[kv.Key]kv.Value {
+	m := map[kv.Key]kv.Value{}
 	for _, kv := range kvs {
 		m[kv.Key] = kv.Value
 	}