+++ api/trace/testtrace/tracer.go
@@ -19,8 +19,6 @@ import (
 	"sync"
 	"time"
 
-	"go.opentelemetry.io/otel/api/kv/value"
-
 	"go.opentelemetry.io/otel/api/kv"
 	"go.opentelemetry.io/otel/api/trace"
 
@@ -82,7 +80,7 @@ func (t *Tracer) Start(ctx context.Context, name string, opts ...trace.StartOpti
 			SpanID:  spanID,
 		},
 		parentSpanID: parentSpanID,
-		attributes:   make(map[kv.Key]value.Value),
+		attributes:   make(map[kv.Key]kv.Value),
 		links:        make(map[trace.SpanContext][]kv.KeyValue),
 	}